requirements.txt
aiogram>=3.2,<4
fastapi>=0.110
uvicorn[standard]>=0.29
httpx>=0.27
pydantic>=1.10
SQLAlchemy>=2.0
python-dotenv>=1.0
alembic>=1.13
PyJWT>=2.8
